# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for صوتي voice data collection app
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users/{userId}/sessions:
    post:
      operationId: adminAssignSession
      tags: [users]
      summary: Assign a session to a user (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignSessionRequest"
      responses:
        "200":
          description: Session assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /user/sessions/{sessionId}/sentences:
    get:
      operationId: getUserSessionSentences
      tags: [sentences]
      summary: Get sentences for a session assigned to current user
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sentences with recording status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SentenceWithRecording"
        "403":
          description: Session not assigned to user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /user/recordings:
    post:
      operationId: submitRecording
      tags: [recordings]
      summary: Submit a recording
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                sentenceId:
                  type: integer
                sessionId:
                  type: integer
                audio:
                  type: string
                  format: binary
              required:
                - sentenceId
                - sessionId
                - audio
      responses:
        "201":
          description: Recording submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingInfo"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /user/suggestions:
    post:
      operationId: createSuggestion
      tags: [suggestions]
      summary: Suggest a new sentence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSuggestionRequest"
      responses:
        "201":
          description: Suggestion created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestionInfo"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [user, admin]
      required:
        - id
        - username
        - role
...
    AssignSessionRequest:
      type: object
      properties:
        sessionId:
          type: integer
      required:
        - sessionId
...
    RecordingInfo:
      type: object
      properties:
        id:
          type: integer
        sentenceId:
          type: integer
        sessionId:
          type: integer
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - sentenceId
        - sessionId
        - status
        - createdAt
...
    SentenceWithRecording:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        orderIndex:
          type: integer
        recordingId:
          type: ["integer", "null"]
        recordingStatus:
          type: ["string", "null"]
      required:
        - id
        - text
        - orderIndex
...
    SuggestionInfo:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        isDuplicate:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - text
        - isDuplicate
        - createdAt
...
    CreateSuggestionRequest:
      type: object
      properties:
        text:
          type: string
      required:
        - text
...